When using a pickup directory for MailKit, return the eml file GUID
When an email is saved to a pickup directory using MailKitSender, it can be useful to return the GUID values that was used to save the eml file. This is similar to returning an identity or sequence value for a database insert operation. The SendResponse already has a MessageId property used by other senders that can be populated with the GUID.